# 8356949: AArch64: Tighten up template interpreter method entry code