@chadrako 2025-05-20
A simple bench mark shows a slight performance improvement but not as much as was measured on x86_64 
```
Benchmark 1: ./build/linux-aarch64-server-release/images/jdk/bin/java -Xms1g -Xmx1g -XX:+UseSerialGC -Xint -jar JavacBenchApp.jar 1
  
# Before

  Time (mean ± σ):      1.361 s ±  0.003 s    [User: 1.328 s, System: 0.042 s]
  Range (min … max):    1.358 s …  1.365 s    10 runs

# After

  Time (mean ± σ):      1.358 s ±  0.002 s    [User: 1.312 s, System: 0.056 s]
  Range (min … max):    1.355 s …  1.362 s    10 runs
```